django_authentication/requirements.txt
djangorestframework-simplejwt==5.4.0
pip-chill==1.0.3
